% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/assignExecutionGroups.R
\name{assignExecutionGroups}
\alias{assignExecutionGroups}
\title{Categorize runs in a raw results file based on minimization and cov step success}
\usage{
assignExecutionGroups(rawres)
}
\arguments{
\item{rawres}{A raw results data frame following the PsN standard.
              Most importantly containing minimization_successful
              and covariance_step_successful columns.}
}
\description{
Categorize runs in a raw results file based on minimization and cov step success
}
\author{
Henrik Bjugård Nyberg - henrik.b.nyberg@farmbio.uu.se
}

