% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPlotlyObj.R
\name{createPlotlyObj}
\alias{createPlotlyObj}
\title{createPlotlyObj}
\usage{
createPlotlyObj(ofvVector, xParamVals, yParamVals,
  paramsToCompare = c("Param1", "Param2"), zlab = "Z: OFV",
  plotTitle = "OFV Surface")
}
\arguments{
\item{ofvVector}{Vector of OFV values for the plot.}

\item{xParamVals}{Vector of X axis values.}

\item{yParamVals}{Vector of Y axis values.}

\item{paramsToCompare}{A vector of two parameters to plot against.}

\item{zlab}{Label for Z axis.}

\item{plotTitle}{Plot title.}
}
\description{
createPlotlyObj
}

