% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileSysSetup.R
\name{fileSysSetup}
\alias{fileSysSetup}
\title{Set up directories for a set of executions}
\usage{
fileSysSetup(modFilePath, runType = "", subDirs = c())
}
\arguments{
\item{modFilePath}{The model file Path. No default.}

\item{runType}{The type of run to be executed. This string is used
in the folder name cerated. Default is an empty string.}

\item{subDirs}{additional subdirectories to create under the main one.
Default is an empty vector.}
}
\description{
Set up directories for a set of executions
}
\author{
Henrik Bjugård Nyberg - henrik.b.nyberg@farmbio.uu.se
}

