% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findRawres.R
\name{findRawres}
\alias{findRawres}
\title{Find the raw results csvs in a specified location}
\usage{
findRawres(path)
}
\arguments{
\item{path}{The path in which to look.}
}
\description{
Find the raw results csvs in a specified location
}
\author{
Henrik Bjugård Nyberg - henrik.b.nyberg@farmbio.uu.se
}

