% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmSaddleWorkflow.R
\name{nmSaddleWorkflow}
\alias{nmSaddleWorkflow}
\title{Full workflow with parallel retries, result categorization and dismount/precond run}
\usage{
nmSaddleWorkflow(modFileName, retries = 9, doParaRetries = TRUE,
  doDismount = FALSE, doNmDismount = TRUE, doCompParaRetries = TRUE,
  rerunDirName, degree = 0.99, seed = 20170115,
  dismountPertDirections = 1, saddleReset = 0, saddleHess = 0,
  slurm_partition = "standard")
}
\arguments{
\item{modFileName}{The model file to use. Default is "run83.mod".}

\item{retries}{The number of retries to run as part of the initial parallel 
retries.}

\item{doParaRetries}{Whether or not the parallel retries should be run. 
Default is TRUE. If FALSE then a rerunDirName is expected.}

\item{doDismount}{Whether or not dismount should be run on all samples. 
Default is TRUE.}

\item{rerunDirName}{If doParaRetries is FALSE, then this directory will be 
assumed to contain a parallel retries run with model files and outputs to 
use for dismount and/or precond. No default.}

\item{dismountPertDirections}{Perturbation direction.}
}
\description{
Full workflow with parallel retries, result categorization and dismount/precond run
}
\author{
Henrik Bjugård Nyberg - henrik.b.nyberg@farmbio.uu.se
}

