% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseExtFile.R
\name{parseExtFile}
\alias{parseExtFile}
\title{parseExtFile}
\usage{
parseExtFile(extFilePath)
}
\arguments{
\item{extFilePath}{File Path to the ext file to be parsed.}
}
\description{
Parses an ext file and return a list of data frame. Each data frame contains the outputs from one estimation step.
}

