% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseNMDataFileName.R
\name{parseNMDataFileName}
\alias{parseNMDataFileName}
\title{Parse the data file name from a NONMEM model}
\usage{
parseNMDataFileName(modFilePath)
}
\arguments{
\item{modFilePath}{The model file Path.


parseNMDataFileName()}
}
\description{
Reads the data file name from a NONMEM control file.
}

