% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseNMMat.R
\name{parseNMMat}
\alias{parseNMMat}
\title{parseNMMat}
\usage{
parseNMMat(nmMatFilePath)
}
\arguments{
\item{nmMatFilePath}{File Path to the ext file to be parsed.}
}
\description{
Parses a matrix as output by NONMEM and return a matrix.
}

