% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseRawresOfvs.R
\name{parseRawresOfvs}
\alias{parseRawresOfvs}
\title{parseRawresOfvs}
\usage{
parseRawresOfvs(rawresPath)
}
\arguments{
\item{rawresPath}{Path to the raw results file to parse OFV values from.}
}
\description{
parseRawresOfvs
}

