% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSurface.R
\name{plotSurface}
\alias{plotSurface}
\title{plotSurface}
\usage{
plotSurface(plotlyAccount, modFilePath, paramsToCompare = c("THETA1",
  "THETA2"), resol = 10, local = FALSE, ...)
}
\arguments{
\item{plotlyAccount}{A username and key for Plotly access on the form plotly(username='user', key='key').}

\item{modFilePath}{Model file to use. The called function createRawresInput assumes that there is an ext file with the same base file name.}

\item{paramsToCompare}{A vector of two parameter names following the NONMEM ext file standard names. Default is c("THETA1", "THETA2"). Parameters to compare. Model file cannot currently use labels for parameters.}

\item{resol}{Resolution on each axis. Default is 10 and will use 10^2 = 100 sets of parameter values, NONMEM runs, and ofv values to create the plot.}

\item{...}{Further options to createRawresInput}
}
\description{
Plots a surface using Plotly and returns the URL
}

