% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runParaRetries.R
\name{runParaRetries}
\alias{runParaRetries}
\title{Run parallel retries}
\usage{
runParaRetries(modFileName, paraRetriesCmd = "parallel_retries",
  dir = paste0("para_retries_", format(Sys.time(), "\%y\%m\%d_\%H\%M\%S")),
  clean = 2, threads = 100, min_retries = 1, degree = 0.1,
  slurm_partition = "standard", rawres_input = "",
  seed = format(Sys.time(), "\%Y\%m\%d"), nm_output = NULL, wait = TRUE)
}
\arguments{
\item{modFileName}{The model to run parallel retries on. Only one permitted.}

\item{paraRetriesCmd}{Command or path used to run parallel retries. Default is "parallel_retries" which assumes the PsN directory is on the PATH.}

\item{dir}{The directory to run in. Defaults to "para_retries_<date>_<time>".}

\item{clean}{The PsN clean-up level. Exactly wrapping the PsN option. Defaults to 2.}

\item{threads}{The number of threads to use. Exactly wrapping the PsN option. Defaults to 100.}

\item{min_retries}{The number of retries to run. Exactly wrapping the PsN option. Defaults to 1

runParaRetries()}
}
\description{
This is a wrapper function for parallel_retries in PsN
}

