% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPsnClean.R
\name{runPsnClean}
\alias{runPsnClean}
\title{Run psn_clean}
\usage{
runPsnClean(path = ".", level = 3)
}
\arguments{
\item{level}{How much to clean up. Higher setting removes more. 2 - 4 is available with 3 being the default.}
}
\description{
Runs the PsN cleanup script. Currently only the level option is mapped and wrapped.
}

