% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stripZeroRowsCols.R
\name{stripZeroRowsCols}
\alias{stripZeroRowsCols}
\title{stripZeroRowsCols}
\usage{
stripZeroRowsCols(inputMatrix)
}
\arguments{
\item{inputMatrix}{Input matrix.}
}
\description{
removes zero rows and columns from a matrix return a matrix.
}

