% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/dismountWorkflow.R
\name{dismountWorkflow}
\alias{dismountWorkflow}
\title{Full workflow with parallel retries, result categorization and dismount/precond run}
\usage{
dismountWorkflow(modFileName = "run83.mod", retries = 9, ...)
}
\arguments{
\item{modFileName}{The model file to use. Default is "run83.mod".}

\item{retries}{The number of retries to run as part of the initial parallel retries.}
}
\description{
Full workflow with parallel retries, result categorization and dismount/precond run
}
\author{
Henrik Bjugård Nyberg - henrik.b.nyberg@farmbio.uu.se
}

