% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUserSlurmQ.R
\name{getUserSlurmQ}
\alias{getUserSlurmQ}
\title{Get SLURM queue for active user}
\usage{
getUserSlurmQ()
}
\description{
Gets the slurm queue for the user that is currently logged in.
No options available. Returns a character string with the rows
of the queue, including the header.
}
\details{
getUserSlurmQ()
}

