% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/parseRawres.R
\name{parseRawres}
\alias{parseRawres}
\title{Parse a PsN raw results file}
\usage{
parseRawres(rawresPath)
}
\arguments{
\item{rawresPath}{The file path for the rawres file to parse. No default.


parseRawres()}
}
\description{
Parse a PsN raw results file
}

