% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/parseRawres.R
\name{parseRawres}
\alias{parseRawres}
\title{Parse a PsN raw results file}
\usage{
parseRawres(rawresPath, addPath = TRUE, addGroup = TRUE)
}
\arguments{
\item{rawresPath}{The file path for the rawres file to parse. No default.}

\item{addPath}{If set the the returned data frame will include a column containing
the file name that was parsed. Default is TRUE.}

\item{addGroup}{If set the returned data frame will include a group column with
       a group number based on minimization and covariance step success. Default
       is TRUE

parseRawres()}
}
\description{
Parse a PsN raw results file
}
\author{
Henrik Bjugård Nyberg - henrik.b.nyberg@farmbio.uu.se
}

