% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseRawres.R
\name{parseRawres}
\alias{parseRawres}
\title{Parse a PsN raw results file}
\usage{
parseRawres(rawresPath, cols = NULL, skipRows = 0, addPath = FALSE,
  addGroup = FALSE)
}
\arguments{
\item{rawresPath}{The file path for the rawres file to parse. No default.}

\item{cols}{A vector of column names or numbers to be parsed. If set to vector of length 1 or longer all other columns will be discarded. Please observe that column names must match the read.csv interpretation of the column header.}

\item{skipRows}{Number of rows to skip after the header. Not to be confused with 'skip' which is a read.csv option that will skip rows from the top of the file before looking for a header.}

\item{addPath}{If set the the returned data frame will include a column containing the file name that was parsed. Default is FALSE.}

\item{addGroup}{If set the returned data frame will include a group column with a group number based on minimization and covariance step success. Default is FALSE

parseRawres()}
}
\description{
Parse a PsN raw results file
}
\author{
Henrik Bjugård Nyberg - henrik.b.nyberg@farmbio.uu.se
}

