% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSurface.R
\name{plotSurface}
\alias{plotSurface}
\title{plotSurface}
\usage{
plotSurface(plotlyAccount, modFilePath, paramsToCompare = c("THETA1",
  "THETA2"), xLims, yLims, resol = 10, local = FALSE)
}
\arguments{
\item{plotlyAccount}{A username and key for Plotly access on the form plotly(username='user', key='key').}

\item{modFilePath}{Model file to use. The called function createRawresInput assumes that there is an ext file with the same base file name.}

\item{xLims}{A vector of two values, lower and upper limit for the first paramsToCompare. Default is 1% below and above the original value}

\item{yLims}{A vector of two values, lower and upper limit for the first paramsToCompare. Default is 1% below and above the original value}

\item{resol}{Resolution on each axis. Default is 10 and will use 10^2 = 100 sets of parameter values, NONMEM runs, and ofv values to create the plot.}
}
\description{
Plots a surface using Plotly and returns the URL
}

