% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSurface.R
\name{plotSurface}
\alias{plotSurface}
\title{plotSurface}
\usage{
plotSurface(plotlyUsername, plotlyKey, modFilePath,
  paramsToCompare = c("Param1", "Param2"), resol = 10, ofvScaling = FALSE,
  absolute = FALSE, slurm_partition = "standard", cleanLevel = 4,
  plotOrigVals = FALSE, ...)
}
\arguments{
\item{plotlyUsername}{Plotly online user name. Not needed if environment variables already set. See plotly instructions.}

\item{plotlyKey}{Plotly online key. Not needed if environment variables already set. See plotly instructions.}

\item{modFilePath}{Model file to use. The called function createRawresInput assumes that there is an ext file with the same base file name.}

\item{paramsToCompare}{Parameters to compare. A vector of two parameter names following the NONMEM ext file standard names. Default is c("THETA1", "THETA2"). Model file parameter labels will be removed.}

\item{resol}{Resolution on each axis. Default is 10 and will use 10^2 = 100 sets of parameter values, NONMEM runs, and ofv values to create the plot.}

\item{ofvScaling}{If true OFVs are scaled to between zero and one. Default is FALSE.}

\item{absolute}{Whether or not to take the absolute value of all parameter values before constructing the input file. Default is FALSE.}

\item{cleanLevel}{PsN clean script will be run on the parallel retries folder. See psn_clean documentation. Default is 4, the highest cleaning level.}

\item{...}{Further options to createRawresInput}

\item{origVals}{Whether or not to plot the original model final estimate as a point in the plot.}
}
\description{
Plots a surface using Plotly and returns the URL
}

