% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDismount.R
\name{runDismount}
\alias{runDismount}
\title{Run dismount (isestimable)}
\usage{
runDismount(modelFileName, wait = FALSE, logging = FALSE,
  runOnSlurm = FALSE)
}
\arguments{
\item{modelFileName}{The name of the model file. No default}

\item{wait}{Wraps the wait option in system(), which specifies whether to
wait for the system call to complete before continuing in R (TRUE) or not
wait after submission (FALSE). Default is FALSE}

\item{logging}{Whether to log the command line outputs to a text file or not. Default is FALSE.

runDismount()}
}
\description{
Runs dismount (aka isestimable). Currently limited to slurm queue submission.
This function is in need of some TLC.
}
\author{
Henrik Bjugård Nyberg - henrik.b.nyberg@farmbio.uu.se
}

