% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/runPrecond.R
\name{runPrecond}
\alias{runPrecond}
\title{Run precond}
\usage{
runPrecond(modelFileName, precondScriptPath = "precond",
  preCondMatFileName = NULL, always = TRUE)
}
\arguments{
\item{modelFileName}{The name of the model file. No default.}

\item{precondScriptPath}{Optional precond script path. Default is precond.}
}
\description{
Run precond
}

