% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPsnClean.R
\name{runPsnClean}
\alias{runPsnClean}
\title{Run psn_clean}
\usage{
runPsnClean(path = ".", level = 3, interact = FALSE, addOpt = "")
}
\arguments{
\item{level}{How much to clean up. Higher setting removes more. 2 - 4 is available with 3 being the default.}

\item{interact}{If TRUE the user is prompted for deletion or every file in an interactive session. Default is FALSE.}

\item{addOpt}{String of additional options to include on command line. Must follow PsN conventions and include space in between options. Default is an empty string.

runPsnClean()}
}
\description{
Runs the PsN cleanup script. Currently only the level option is mapped and wrapped.
}

